removed compiler warnings

+++ opm/simulators/wells/WellHelpers.hpp
@@ -238,8 +238,8 @@ namespace Opm {
         DenseMatrix transposeDenseDynMatrix(const DenseMatrix& M)
         {
             DenseMatrix tmp{M.cols(), M.rows()};
-            for (int i = 0; i < M.rows(); ++i) {
-                for (int j = 0; j < M.cols(); ++j) {
+            for (size_t i = 0; i < M.rows(); ++i) {
+                for (size_t j = 0; j < M.cols(); ++j) {
                     tmp[j][i] = M[i][j];
                 }
             }
